Made many objects message able
Created the MessageReceiver class to use in the different Chat Types and the User Class. This class is used to allow sending messages to entities. This was previously described in #1.
Added the IllegalChatTypeException to throw if the wrong #getAs... Method is called from the GenericChat Class. This replaces the previously thrown IllegalStateException.