added the basic update and event logic
Added the continuous update cycle to check for updates.
Added Events which can be fired by such Updates.
Added the UpdateParser to parse incoming update-json Strings to updates.
Added the EventManager in the JTA and in the JTABuilder.
Added the EventListener and EventHandler to allow dynamic listening to events.
Added the ChatMessageReceivedEvent without content for testing. Might get renamed to MessageReceivedEvent and filled with its corresponding content.