New Chat Functionality
Added LeaveChatRequest which is called in GenericChat#leave
Added the SendMessageAction, as mentioned in #5.
Modified the IllegalChatTypeException to allow own messages, to throw a specific Error in PrivateChat#leave.
Added multiple functions to modify the SendMessageRequest.
Added Support for the GroupChat Object, which is now finally parsed from its Json.
Modified Message#reply to use SendMessageAction